# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "servers": [],
  "info": {
    "title": "FutHub API",
    "version": "1.0",
    "description": "API do FutHub para jogo, automações Discord e administração."
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
  "paths": {
...
    "/v1/club": {
      "post": {
        "tags": ["Clube"],
        "operationId": "getClub",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": ["id", "name", "avatarUrl"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "balance": {
                      "type": "number"
                    },
                    "stadium": {
                      "additionalProperties": false,
                      "properties": {
                        "level": {
                          "type": "number"
                        },
                        "maxLevel": {
                          "type": "number"
                        },
                        "nextUpgradeCost": {
                          "type": "number",
                          "nullable": true
                        },
                        "ticketRevenue": {
                          "type": "number"
                        },
                        "maintenance": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "level",
                        "maxLevel",
                        "nextUpgradeCost",
                        "ticketRevenue",
                        "maintenance"
                      ],
                      "type": "object"
                    },
                    "sponsor": {
                      "additionalProperties": false,
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "weeklyMatches": {
                          "type": "number"
                        },
                        "weeklyGoal": {
                          "type": "number"
                        },
                        "payout": {
                          "type": "number"
                        },
                        "completed": {
                          "type": "boolean"
                        }
                      },
                      "required": ["name", "weeklyMatches", "weeklyGoal", "payout", "completed"],
                      "type": "object"
                    },
                    "payroll": {
                      "type": "number"
                    },
                    "projectedNet": {
                      "type": "number"
                    }
                  },
                  "required": ["balance", "stadium", "sponsor", "payroll", "projectedNet"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/v1/club/stadium/upgrade": {
      "post": {
        "tags": ["Clube"],
        "operationId": "upgradeStadium",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "nullable": true
                  }
                },
                "required": ["id", "name", "avatarUrl"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "balance": {
                      "type": "number"
                    },
                    "stadium": {
                      "additionalProperties": false,
                      "properties": {
                        "level": {
                          "type": "number"
                        },
                        "maxLevel": {
                          "type": "number"
                        },
                        "nextUpgradeCost": {
                          "type": "number",
                          "nullable": true
                        },
                        "ticketRevenue": {
                          "type": "number"
                        },
                        "maintenance": {
                          "type": "number"
                        }
                      },
                      "required": [
                        "level",
                        "maxLevel",
                        "nextUpgradeCost",
                        "ticketRevenue",
                        "maintenance"
                      ],
                      "type": "object"
                    },
                    "sponsor": {
                      "additionalProperties": false,
                      "properties": {
                        "name": {
                          "type": "string"
                        },
                        "weeklyMatches": {
                          "type": "number"
                        },
                        "weeklyGoal": {
                          "type": "number"
                        },
                        "payout": {
                          "type": "number"
                        },
                        "completed": {
                          "type": "boolean"
                        }
                      },
                      "required": ["name", "weeklyMatches", "weeklyGoal", "payout", "completed"],
                      "type": "object"
                    },
                    "payroll": {
                      "type": "number"
                    },
                    "projectedNet": {
                      "type": "number"
                    }
                  },
                  "required": ["balance", "stadium", "sponsor", "payroll", "projectedNet"],
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
...
}